* using eval on diagsman

+++ Plugins/Diags/Source/Diags/Diags.h
@@ -93,7 +93,7 @@ public:
 	//	"sqrt({myflag}^2)" absolute value of myflag (root of a squared number)
 	// Res is the result of the evaluation (a number)
 	UFUNCTION(BlueprintCallable, BlueprintPure=false, Category="Dialogs")
-	bool CheckCondition(const FString& Expression, float& Res) const;
+	bool CheckCondition(const FString& Expression, double& Res) const;
 
 #pragma region Init
 	// set the data to be used. call upon initialization.
